requirements.txt
django==5.1.1
djangorestframework==3.15.2
pytest-django==4.9.0
model_bakery==1.19.5
Pillow==11.0.0
django-cleanup==9.0.0
django-cors-headers==4.6.0